Update social-pressing tool

+++ .innoobwetrust/life_hacks/social-pressing/src/utils.rs
@@ -4,7 +4,7 @@ use std::thread;
 use tracing::{debug, instrument};
 
 pub(crate) fn rand_delay_duration() -> Duration {
-    let millis = (rand::random::<usize>() % 3 + 2) * 1000 + (rand::random::<usize>() % 1000 + 1); // Max 5s
+    let millis = (rand::random::<u64>() % 3 + 2) * 1000 + (rand::random::<u64>() % 1000 + 1); // Max 5s
     let millis = Duration::from_millis(millis as u64);
 
     millis
